requirements.txt
pytest
pytest-bdd
pytest-html
pyyaml
selenium